+++ server/ClientsCollection.py
@@ -1,0 +1,5 @@
+import collections
+
+class ClientsCollection(collections.UserDict):
+    def sockets(self):
+        return list(self.data.keys())
