+++ server/__init__.py
@@ -1,0 +1,7 @@
+from state import *
+from SMTP_FSM import *
+from ClientSocket import *
+from Client import *
+from MailServer import *
+from ClientsCollection import *
+from default import *
