--- server/config.ini
@@ -1,2 +1,0 @@
-SERVER_DOMAIN=superserver.local
-SERVER_PORT=2556
