+++ server/default.py
@@ -1,0 +1,6 @@
+SERVER_DOMAIN = 'superserver.local'
+SERVER_PORT   = 2556
+READ_TIMEOUT  = 40
+
+DEFAULT_SUPR_DIR = f'../pst/local/'
+DEFAULT_USER_DIR = f'../pst/maildir/'
